user domain integration setup
