+++ src/User/composer.json
@@ -34,7 +34,8 @@
         "symfony/var-dumper": "^3.4|^4.0"
     },
     "suggest": {
-        "msgphp/eav": "To enable user attribute values"
+        "msgphp/eav": "To enable user attribute values",
+        "msgphp/organization": "To enable user organizations"
     },
     "config": {
         "preferred-install": {
